Make Context type-generic and unify parsing funcs
Converted Context into a generic form that is instantiated with the
different LC types.
Remove much of the duplication between Pure.Parser and
SimpleTypes.Parser into the new Parsing module and re-wrote these
parsers using the new typeclasses.  These will likely change as
more calculi are added, but serve the same purpose as the LambdaTerm
dataype in reducing the large amounts of duplication present in the
project.