Rearrange some Parsing instances
To fix the orphan warnings, I moved the instances to the same file as
the typeclass declarations.  This caused some other rearranging,
including deleting the Pure module and putting the PureTerm synonym
in LambdaTerm and moving the locallyClosedCheck function to LambdaTerm
as well.

Additionally, I implemented a couple new relevant typeclasses for Nat
to better align with standard library function names rather than
forcing the need to use Nat-specific ones.